Extract two factor code from __init__
