+++ cmk/gui/userdb/__init__.py
@@ -93,6 +93,11 @@ from ._check_credentials import (
 from ._check_credentials import user_exists as user_exists
 from ._check_credentials import user_exists_according_to_profile as user_exists_according_to_profile
 from ._check_credentials import user_locked as user_locked
+from ._two_factor import disable_two_factor_authentication as disable_two_factor_authentication
+from ._two_factor import is_two_factor_backup_code_valid as is_two_factor_backup_code_valid
+from ._two_factor import is_two_factor_login_enabled as is_two_factor_login_enabled
+from ._two_factor import load_two_factor_credentials as load_two_factor_credentials
+from ._two_factor import make_two_factor_backup_codes as make_two_factor_backup_codes
 from ._user_attribute import register_custom_user_attributes as register_custom_user_attributes
 from ._user_attribute import (
     update_config_based_user_attributes as update_config_based_user_attributes,
@@ -228,69 +233,6 @@ def need_to_change_pw(username: UserId, now: datetime) -> str | None:
     if now.timestamp() - last_pw_change > max_pw_age:
         return "expired"
     return None
-
-
-def is_two_factor_login_enabled(user_id: UserId) -> bool:
-    """Whether or not 2FA is enabled for the given user"""
-    return bool(
-        load_two_factor_credentials(user_id)["webauthn_credentials"]
-        or load_two_factor_credentials(user_id)["totp_credentials"]
-    )
-
-
-def disable_two_factor_authentication(user_id: UserId) -> None:
-    credentials = load_two_factor_credentials(user_id, lock=True)
-    credentials["webauthn_credentials"].clear()
-    credentials["totp_credentials"].clear()
-    save_two_factor_credentials(user_id, credentials)
-
-
-def load_two_factor_credentials(user_id: UserId, lock: bool = False) -> TwoFactorCredentials:
-    cred = load_custom_attr(
-        user_id=user_id, key="two_factor_credentials", parser=ast.literal_eval, lock=lock
-    )
-    return (
-        TwoFactorCredentials(webauthn_credentials={}, backup_codes=[], totp_credentials={})
-        if cred is None
-        else cred
-    )
-
-
-def make_two_factor_backup_codes(
-    *, rounds: int | None = None
-) -> list[tuple[Password, PasswordHash]]:
-    """Creates a set of new two factor backup codes
-
-    The codes are returned in plain form for displaying and in hashed+salted form for storage
-    """
-    return [
-        (password, password_hashing.hash_password(password))
-        for password in (Password.random(10) for _ in range(10))
-    ]
-
-
-def is_two_factor_backup_code_valid(user_id: UserId, code: Password) -> bool:
-    """Verifies whether or not the given backup code is valid and invalidates the code"""
-    credentials = load_two_factor_credentials(user_id)
-    matched_code = None
-
-    for stored_code in credentials["backup_codes"]:
-        try:
-            password_hashing.verify(code, stored_code)
-            matched_code = stored_code
-            break
-        except (password_hashing.PasswordInvalidError, ValueError):
-            continue
-
-    if matched_code is None:
-        return False
-
-    # Invalidate the just used code
-    credentials = load_two_factor_credentials(user_id, lock=True)
-    credentials["backup_codes"].remove(matched_code)
-    save_two_factor_credentials(user_id, credentials)
-
-    return True
 
 
 def _is_local_user(user: UserSpec) -> bool:
